A4/streamlit/requirements.txt
Flask
Flask-JWT-Extended
psycopg2-binary
streamlit
seaborn
matplotlib
pandas
